+++ client/src/components/OurPartners/OurPartnersConnections.jsx
