+++ client/src/components/OurPartners/OurPartnersHeader.jsx
